% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.metadata.R
\docType{data}
\name{dom.metadata}
\alias{dom.metadata}
\title{metadata
Data frame with metadata associated with each matrix}
\format{Data frame of 15 
  \describe{
    \item{matrix}{List of matrices of dominance interactions}
    \item{fileid}{File identifier associated with each matrix. Format is Author_Year}
    \item{taxon}{Broad taxonomic category for species}
    \item{and so on}{...}
 }}
\usage{
dom.metadata
}
\description{
metadata
Data frame with metadata associated with each matrix
}
\keyword{datasets}
