% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_archive.R
\name{subset_archive}
\alias{subset_archive}
\title{Subset DomArchive data}
\usage{
subset_archive(
  dataset = dom.data,
  fileids = NULL,
  columns = NULL,
  values = NULL,
  return.intxdata.only = FALSE
)
}
\arguments{
\item{dataset}{The dataset to be subsetted. Defaults to dom.data, the data 
object supplied with the package. Only objects with the same structure as dom.data
will work.}

\item{fileids}{Optional set of unique identifiers to be used to subset `dataset`. These must 
be in names(dataset). If no fileids supplied, filtering will be done using `columns` and `values`.}

\item{columns}{Only used if fileids is NULL. A list of column names (must be in `names(dom.metadata)`)
to filter data by.}

\item{values}{Only used if fileids is NULL. A list of values corresponding to
the columns supplied in `columns`. Only rows with `values` at `columns` will
be retained. Must be supplied with `columns` and must be the same length as `columns`.}

\item{return.intxdata.only}{Should only matrices be returned (TRUE), or should both
matrix and metadata be returned (FALSE).}
}
\description{
Select certain data or data + metadata from `dom.data` by filtering by columns
in the metadata or by supplying a list of fileids.
}
\examples{

## Data from SchjelderupEbbe's 1922 paper on pecking order of domestic hens
subset_archive(fileids = c('SchjelderupEbbe_1922a', 'SchjelderupEbbe_1922b'))

## All data from captive primates
subset_archive(columns = list('taxon', 'captivity'), values = list('Primate', 'Captive'))

#' ## All data from captive primates - subset using dom.metadata
subset_archive(fileids = dom.metadata[dom.metadata$taxon == 'Primate' & dom.metadata$captivity == 'Captive',]$fileid)

}
