% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.data.R
\docType{data}
\name{dom.data}
\alias{dom.data}
\title{dom.data}
\format{List of 16 elements
  \describe{
    \item{matrix}{List of matrices of dominance interactions}
    \item{fileid}{File identifier associated with each matrix. Format is Author_Year}
    \item{taxon}{Broad taxonomic category for species}
    \item{and so on}{...}
  }}
\usage{
dom.data
}
\description{
List of dominance matrices and associated metadata. List is organized such that
the 1st dominance matrix corresponds to the 1st entry in the other elements of the list,
and so on.
}
\keyword{datasets}
