% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom.metadata.R
\docType{data}
\name{dom.metadata}
\alias{dom.metadata}
\title{Metadata dataframe}
\format{
Data frame of metadata with one row per dataset in the archive.
Data frame includes biological metadata and summary statistics associated with
each dataset. See 'introduction' vignette for more details.
}
\usage{
dom.metadata
}
\description{
A dataframe containing the metadata for the complete dataset. 
The information in the \code{dom.metadata} dataframe is exactly identical to what you would get if you
extracted and combined every \code{metadata} element in \code{dom.data}.
}
\keyword{datasets}
